pyproject.toml
[project]
name = "qwen-vl-series-finetune"
version = "0.1.0"
description = "Fine-tuning framework for Qwen-VL Series models"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "accelerate==1.10.1",
    "av==14.4.0",
    "bitsandbytes==0.49.2",
    "datasets==3.5.1",
    "decord==0.6.0",
    "deepspeed==0.17.5",
    "einops==0.8.1",
    "huggingface-hub==1.6.0",
    "liger_kernel==0.7.0",
    "numpy==2.1.2",
    "opencv-python==4.11.0.86",
    "pandas==2.2.3",
    "peft==0.15.2",
    "pillow==11.3.0",
    "pydantic==2.11.3",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "tensorboardX==2.6.2.2",
    "tokenizers==0.22.2",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "torchvision==0.23.0",
    "tqdm==4.67.1",
    "transformers==5.3.0",
    "trl==0.25.0",
    "ujson==5.10.0",
    "wandb==0.19.10",
    "qwen-vl-utils",
    "gradio",
]

[tool.uv]
find-links = ["https://download.pytorch.org/whl/cu128"]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
